requirements.txt
numpy>=2.0
scipy>=1.14
pandas>=2.2
web3>=6.0
eth-account>=0.11
plotly>=6.0
streamlit>=1.35
websocket-client>=1.7
requests>=2.31
aiohttp>=3.9
pyyaml>=6.0
python-dotenv>=1.0
